requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-decouple==2.0
django-filter==21.1
django-mongoengine==0.4.4
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djongo==1.3.6
dnspython==2.2.1
drf-excel==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
Markdown==3.3.6
MarkupSafe==2.1.1
mongoengine==0.24.0
numpy==1.22.3
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
psycopg2==2.9.3
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.7.0
